assignment_2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
Pillow>=9.5.0
pycocotools>=2.0.6
